+++ presenter-backstack-nav3/public/src/commonMain/kotlin/software/ralf/app/platform/presenter/backstack/nav3/PresenterBackstackScope.kt
@@ -134,9 +134,7 @@ public fun CompositionLocal<PresenterBackstackScope?>.requireNotNull(): Presente
  * ) : ComposeRenderer<WelcomePresenter.Model>() {
  *   @Composable
  *   override fun Compose(model: WelcomePresenter.Model, modifier: Modifier) {
- *     rendererFactory
- *       .getComposeRenderer(model.contentModel)
- *       .renderCompose(model.contentModel, modifier)
+ *     rendererFactory.renderCompose(model.contentModel, modifier)
  *   }
  * }
  * ```