+++ recipes/common/impl/src/commonMain/kotlin/software/ralf/app/platform/recipes/backstack/CrossSlideBackstackRenderer.kt
@@ -18,7 +18,7 @@ import software.ralf.app.platform.inject.ContributesRenderer
 import software.ralf.app.platform.presenter.BaseModel
 import software.ralf.app.platform.presenter.backstack.nav3.PresenterBackstackRenderer
 import software.ralf.app.platform.renderer.RendererFactory
-import software.ralf.app.platform.renderer.getComposeRenderer
+import software.ralf.app.platform.renderer.renderCompose
 
 @Inject
 @ContributesRenderer
@@ -48,7 +48,7 @@ class CrossSlideBackstackRenderer(private val rendererFactory: RendererFactory) 
 
   @Composable
   override fun ComposeBackstackEntry(model: BaseModel) {
-    rendererFactory.getComposeRenderer(model).renderCompose(model)
+    rendererFactory.renderCompose(model)
   }
 
   private fun AnimatedContentTransitionScope<Scene<Int>>.crossSlideTransition(